Consolidate GitHub embed trio into <GitHubContent>
Collapses GitHubPage (230 LOC), GitHubPR (468 LOC), and
GitHubPageWithPR (76 LOC) into a single GitHubContent component
(693 LOC including both internal views). Net -736 LOC across the
component layer; every callsite's prop list also shrinks.

Public API:
  <GitHubContent
    repository="org/repo"
    filePath="path/to/file.ext"
    branch?       // default "main"
    pr?={{ number, focusFile? }}    // opt in to PR diff tab
    title? description? className?
  />

When `pr` is present, renders the tabbed Final Implementation /
GitHub Changes UI; absent, renders the file viewer only. The tabbed
mode only mounts the active view so PR diff fetches still don't
fire until the user clicks the diff tab (matches the old
GitHubPageWithPR ternary).

Callsite migration covered all 13 usages across triggers,
drive-to-point, advanced-drive-to-point, logging-implementation,
pathplanner, vision-shooting, vision-implementation, and
MechanismTabs. Docs updated in CLAUDE.md, README.md,
.github/CONTRIBUTING.md, and the source comment in githubSchemas.ts.
MODERNIZATION_STATUS.md marks the queued audit item as shipped with
the full session log appended.

Validated: pnpm type-check + pnpm lint clean. Visual smoke at
/drive-to-point renders the file viewer + walkthrough chain without
regression against the engineering reskin earlier on this branch.
